forge fmt

+++ test/utils/Moon.sol
@@ -3,16 +3,14 @@ pragma solidity ^0.8.7;
 
 import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 
-contract Moon is ERC20{
-
-    constructor (string memory _name, string memory _symbol) ERC20 (_name,_symbol){
-    }
+contract Moon is ERC20 {
+    constructor(string memory _name, string memory _symbol) ERC20(_name, _symbol) { }
 
     function mint(address to, uint256 amount) public virtual {
-        _mint(to,amount);
+        _mint(to, amount);
     }
 
-    function burn(address form, uint amount) public virtual {
+    function burn(address form, uint256 amount) public virtual {
         _burn(form, amount);
     }
 }
